backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google_search_results==2.4.2
groq>=0.13.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5